+++ fe-vm/src/components/main/Info.jsx
