+++ fe-vm/src/components/main/VendingMachine.jsx
